@babel/core: parse methods should parse only
