+++ packages/babel-core/src/parse.js
@@ -1,7 +1,7 @@
 // @flow
 
 import loadConfig, { type InputOptions } from "./config";
-import normalizeFile from "./transformation/normalize-file";
+import parser from "./parser";
 import normalizeOptions from "./transformation/normalize-opts";
 
 type AstRoot = BabelNodeFile | BabelNodeProgram;
@@ -49,7 +49,7 @@ export const parse: Parse = (function parse(code, opts, callback) {
       const cfg = loadConfig(opts);
       if (cfg === null) return cb(null, null);
 
-      ast = normalizeFile(cfg.passes, normalizeOptions(cfg), code).ast;
+      ast = parser(cfg.passes, normalizeOptions(cfg), code);
     } catch (err) {
       return cb(err);
     }
@@ -68,7 +68,7 @@ export function parseSync(
     return null;
   }
 
-  return normalizeFile(config.passes, normalizeOptions(config), code).ast;
+  return parser(config.passes, normalizeOptions(config), code);
 }
 
 export function parseAsync(
